% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbit.R
\name{orbit}
\alias{orbit}
\title{Calculate the orbit of a one dimensional map.}
\usage{
orbit(x0, map, T)
}
\arguments{
\item{x0}{Initial value.}

\item{map}{A function with one input that describes the map.}

\item{T}{The number of iterations.}
}
\value{
A vector of the orbit.
}
\description{
Calculate the orbit of a one dimensional map.
}
\examples{
orbit(0.1, function(x) 0.85 * sin(pi * x), 100)
orbit(0.2, function(x) mapLogistic(x, 3.2), 200)
}

