% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCobweb.R
\name{plotCobweb}
\alias{plotCobweb}
\title{Cobweb plot.}
\usage{
plotCobweb(map, x0, n.iter, from = 0, to = 1)
}
\arguments{
\item{map}{A function with one input that describes the map.}

\item{x0}{Initial value.}

\item{n.iter}{The number of iterations.}

\item{from}{The lower bound of the x axis. Defaults to 0.}

\item{to}{The upper bound of the x axis. Defaults to 1.}
}
\value{
None
}
\description{
Cobweb plot.
}
\examples{
plotCobweb(function(x) 1.2 * pmin(x, 1 - x), 0.1, 100) # Tent map
}

