% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOrbit.R
\name{plotOrbit}
\alias{plotOrbit}
\title{Plot an orbit.}
\usage{
plotOrbit(x)
}
\arguments{
\item{x}{A vector of the orbit.}
}
\value{
None
}
\description{
Plot an orbit.
}
\examples{
plotOrbit(orbit(0.1, function(x) 0.89 * sin(pi * x), 100))
}

